+++ connect/api/v2/template_projects/serializers.py
@@ -1,14 +1,6 @@
 from rest_framework.serializers import ModelSerializer
-from connect.template_projects.models import TemplateType, TemplateAI, TemplateFeature
+from connect.template_projects.models import TemplateType, TemplateFeature
 from rest_framework import serializers
-
-
-class TemplateAISerializer(ModelSerializer):
-
-    class Meta:
-
-        model = TemplateAI
-        fields = "__all__"
 
 
 class TemplateFeatureSerializer(ModelSerializer):
@@ -21,21 +13,17 @@ class TemplateFeatureSerializer(ModelSerializer):
 class TemplateTypeSerializer(ModelSerializer):
 
     features = serializers.SerializerMethodField()
-    ais = serializers.SerializerMethodField()
 
     class Meta:
         model = TemplateType
         fields = [
             'id', 'category', 'description', 'name',
             'level', 'setup', 'photo', 'features',
-            'ais', 'photo_description', 'uuid', 'base_project_uuid'
+            'photo_description', 'uuid', 'base_project_uuid'
         ]
 
     def get_features(self, obj):
         return TemplateFeatureSerializer(obj.template_features.all(), many=True).data
-
-    def get_ais(self, obj):
-        return TemplateAISerializer(obj.template_ais.all(), many=True).data
 
 
 class RetrieveTemplateSerializer(ModelSerializer):