+++ connect/api/v2/template_projects/tests.py
@@ -1,9 +1,9 @@
 import json
 from django.test import TestCase, RequestFactory
-from .views import TemplateTypeViewSet, TemplateAIViewSet, TemplateFeatureViewSet
+from .views import TemplateTypeViewSet, TemplateFeatureViewSet
 
 from connect.api.v1.tests.utils import create_user_and_token
-from connect.template_projects.models import TemplateType, TemplateAI, TemplateFeature
+from connect.template_projects.models import TemplateType, TemplateFeature
 
 
 class TemplateTypeViewSetTestCase(TestCase):
@@ -57,47 +57,6 @@ class TemplateTypeViewSetTestCase(TestCase):
             id=template_id
         )
         self.assertEqual(response["name"], "name")
-
-
-class TemplateAIViewSetTest(TestCase):
-
-    def setUp(self):
-
-        self.factory = RequestFactory()
-        self.owner, self.owner_token = create_user_and_token("owner")
-
-        self.template_type_object = TemplateType.objects.create(
-            level=1,
-            category="category",
-            description="description",
-            name="name"
-        )
-
-        self.template_ai_object = TemplateAI.objects.create(
-            name="name",
-            description="description",
-            template_type=self.template_type_object
-        )
-
-    def request(self, method, url, data=None, user=None, token=None, id=None):
-
-        headers = {"HTTP_AUTHORIZATION": "Token {}".format(token.key)} if token else {}
-
-        request = self.factory.request(method=method, path=url, data=data, **headers)
-        response = TemplateAIViewSet.as_view(method)(request, pk=id)
-        response.render()
-        content_data = json.loads(response.content)
-
-        return content_data
-
-    def test_get_queryset(self):
-
-        response = self.request({"get": "list"}, "/v2/projects/template-ai", user=self.owner, token=self.owner_token)
-        self.assertEqual(response["count"], 1)
-        response = self.request({"get": "list"}, "/v2/projects/template-ai?name=name", user=self.owner, token=self.owner_token)
-        self.assertEqual(response["count"], 1)
-        response = self.request({"get": "list"}, "/v2/projects/template-ai?template_type=1", user=self.owner, token=self.owner_token)
-        self.assertEqual(response["count"], 1)
 
 
 class TemplateFeatureViewSetTest(TestCase):