+++ connect/api/v2/user/tests.py
@@ -1,4 +1,3 @@
-
 import uuid as uuid4
 from unittest.mock import Mock, patch
 from rest_framework import status